Implemented CustomTabBar to switch between non-locked and locked notes.

+++ CleanPad/Presentation/ListViews/NonLockedNotesListView.swift
@@ -10,31 +10,46 @@ import SwiftUI
 /// View that contains non-locked notes, and let users tap a note to view and edit it.
 struct NonLockedNotesListView: View {
     @ObservedObject var viewModel: NotesListViewModel
+    @Binding var showEditViewSheet: Bool
     
     var body: some View {
-        List {
-            ForEach(viewModel.nonLockedNotes) { note in
-                NavigationLink {
-                    // Open NoteEditView with the tapped note.
-                    NoteEditView(note: note, viewModel: viewModel, creatingNewNote: false)
-                } label: {
-                    VStack(alignment: .leading) {
-                        Text(note.title)
-                        Text(note.date.formatted(date: .abbreviated, time: .shortened))
-                            .font(.caption)
-                            .foregroundColor(.secondary)
-                    }
-                    .contextMenu {
-                        // Button to change isLocked note property, i.e., move it to the personal space.
-                        Button {
-                            viewModel.updateLockStatus(for: note)
+        if viewModel.nonLockedNotes.isEmpty {
+            EmptyListView(
+                imageSystemName: "note.text",
+                label: "This looks a little empty...",
+                description: viewModel.placeholders.randomElement() ?? "Start writing...",
+                buttonLabel: "Create a note!"
+            ) {
+                showEditViewSheet.toggle()
+            }
+            .padding(.bottom, 80)
+        } else {
+            Form {
+                List {
+                    ForEach(viewModel.nonLockedNotes) { note in
+                        NavigationLink {
+                            // Open NoteEditView with the tapped note.
+                            NoteEditView(note: note, viewModel: viewModel, creatingNewNote: false)
                         } label: {
-                            Label("Move to personal space", systemImage: "lock.open.fill")
+                            VStack(alignment: .leading) {
+                                Text(note.title)
+                                Text(note.date.formatted(date: .abbreviated, time: .shortened))
+                                    .font(.caption)
+                                    .foregroundColor(.secondary)
+                            }
+                            .contextMenu {
+                                // Button to change isLocked note property, i.e., move it to the personal space.
+                                Button {
+                                    viewModel.updateLockStatus(for: note)
+                                } label: {
+                                    Label("Move to personal space", systemImage: "lock.open.fill")
+                                }
+                            }
                         }
                     }
+                    .onDelete(perform: viewModel.removeNonLockedNoteFromList)
                 }
             }
-            .onDelete(perform: viewModel.removeNonLockedNoteFromList)
         }
     }
 }
